moved states to features
